requirements.txt
gTTS==2.5.3
python-dotenv==1.0.1
requests==2.32.3
pydub==0.25.1